refactor: URLSessionDataTask를 만드는 함수 분리
