--- OpenMarket/OpenMarket/Model/SalesInformation.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-struct SalesInformation: Codable {
-    var name: String
-    var descriptions: String
-    var price: Double
-    var currency: Currency
-    var discountedPrice: Double?
-    var stock: Int?
-    var secret: String
-}
